+++ server/base/src/main/java/org/apache/accumulo/server/ServerInfo.java
@@ -53,6 +53,7 @@ public class ServerInfo implements ClientInfo {
   private int zooKeepersSessionTimeOut;
   private VolumeManager volumeManager;
   private ZooCache zooCache;
+  private final ServerConstants serverConstants;
 
   ServerInfo(SiteConfiguration siteConfig, String instanceName, String zooKeepers,
       int zooKeepersSessionTimeOut) {
@@ -82,6 +83,7 @@ public class ServerInfo implements ClientInfo {
       throw new RuntimeException("Instance id " + instanceID + " pointed to by the name "
           + instanceName + " does not exist in zookeeper");
     }
+    serverConstants = new ServerConstants(siteConfig, hadoopConf);
   }
 
   ServerInfo(SiteConfiguration config) {
@@ -93,7 +95,8 @@ public class ServerInfo implements ClientInfo {
     } catch (IOException e) {
       throw new IllegalStateException(e);
     }
-    Path instanceIdPath = ServerUtil.getAccumuloInstanceIdPath(volumeManager);
+    serverConstants = new ServerConstants(siteConfig, hadoopConf);
+    Path instanceIdPath = serverConstants.getInstanceIdLocation(volumeManager.getFirst());
     instanceID = VolumeManager.getInstanceIDFromHdfs(instanceIdPath, hadoopConf);
     zooKeepers = config.get(Property.INSTANCE_ZK_HOST);
     zooKeepersSessionTimeOut = (int) config.getTimeInMillis(Property.INSTANCE_ZK_TIMEOUT);
@@ -115,6 +118,7 @@ public class ServerInfo implements ClientInfo {
     zooKeepersSessionTimeOut = (int) config.getTimeInMillis(Property.INSTANCE_ZK_TIMEOUT);
     zooCache = new ZooCacheFactory().getZooCache(zooKeepers, zooKeepersSessionTimeOut);
     this.instanceName = instanceName;
+    serverConstants = new ServerConstants(siteConfig, hadoopConf);
   }
 
   public SiteConfiguration getSiteConfiguration() {
@@ -178,5 +182,9 @@ public class ServerInfo implements ClientInfo {
   @Override
   public Configuration getHadoopConf() {
     return this.hadoopConf;
+  }
+
+  public ServerConstants getServerConstants() {
+    return serverConstants;
   }
 }
