+++ server/manager/src/main/java/org/apache/accumulo/manager/upgrade/UpgradeCoordinator.java
@@ -18,6 +18,7 @@
  */
 package org.apache.accumulo.manager.upgrade;
 
+import java.io.IOException;
 import java.util.Map;
 import java.util.OptionalInt;
 import java.util.concurrent.CompletableFuture;
@@ -25,12 +26,19 @@ import java.util.concurrent.Future;
 import java.util.concurrent.SynchronousQueue;
 import java.util.concurrent.TimeUnit;
 
+import org.apache.accumulo.core.Constants;
+import org.apache.accumulo.core.client.AccumuloException;
 import org.apache.accumulo.core.dataImpl.KeyExtent;
 import org.apache.accumulo.core.util.threads.ThreadPools;
+import org.apache.accumulo.core.volume.Volume;
+import org.apache.accumulo.fate.ReadOnlyStore;
+import org.apache.accumulo.fate.ReadOnlyTStore;
+import org.apache.accumulo.fate.ZooStore;
 import org.apache.accumulo.manager.EventCoordinator;
 import org.apache.accumulo.server.ServerConstants;
 import org.apache.accumulo.server.ServerContext;
-import org.apache.accumulo.server.ServerUtil;
+import org.apache.accumulo.server.fs.VolumeManager;
+import org.apache.hadoop.fs.Path;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -131,8 +139,9 @@ public class UpgradeCoordinator {
         "Not currently in a suitable state to do zookeeper upgrade %s", status);
 
     try {
-      int cv = ServerUtil.getAccumuloPersistentVersion(context.getVolumeManager());
-      ServerUtil.ensureDataVersionCompatible(cv);
+      int cv = context.getServerConstants()
+          .getAccumuloPersistentVersion(context.getVolumeManager().getFirst());
+      ServerContext.ensureDataVersionCompatible(cv);
       this.currentVersion = cv;
 
       if (cv == ServerConstants.DATA_VERSION) {
@@ -141,7 +150,7 @@ public class UpgradeCoordinator {
       }
 
       if (currentVersion < ServerConstants.DATA_VERSION) {
-        ServerUtil.abortIfFateTransactions(context);
+        abortIfFateTransactions(context);
 
         for (int v = currentVersion; v < ServerConstants.DATA_VERSION; v++) {
           log.info("Upgrading Zookeeper from data version {}", v);
@@ -182,7 +191,8 @@ public class UpgradeCoordinator {
               }
 
               log.info("Updating persistent data version.");
-              ServerUtil.updateAccumuloVersion(context.getVolumeManager(), currentVersion);
+              updateAccumuloVersion(context.getServerConstants(), context.getVolumeManager(),
+                  currentVersion);
               log.info("Upgrade complete");
               setStatus(UpgradeStatus.COMPLETE, eventCoordinator);
             } catch (Exception e) {
@@ -195,7 +205,66 @@ public class UpgradeCoordinator {
     }
   }
 
+  // visible for testing
+  synchronized void updateAccumuloVersion(ServerConstants constants, VolumeManager fs,
+      int oldVersion) {
+    for (Volume volume : fs.getVolumes()) {
+      try {
+        if (constants.getAccumuloPersistentVersion(volume) == oldVersion) {
+          log.debug("Attempting to upgrade {}", volume);
+          Path dataVersionLocation = constants.getDataVersionLocation(volume);
+          fs.create(new Path(dataVersionLocation, Integer.toString(ServerConstants.DATA_VERSION)))
+              .close();
+          // TODO document failure mode & recovery if FS permissions cause above to work and below
+          // to fail ACCUMULO-2596
+          Path prevDataVersionLoc = new Path(dataVersionLocation, Integer.toString(oldVersion));
+          if (!fs.delete(prevDataVersionLoc)) {
+            throw new RuntimeException("Could not delete previous data version location ("
+                + prevDataVersionLoc + ") for " + volume);
+          }
+        }
+      } catch (IOException e) {
+        throw new RuntimeException("Unable to set accumulo version: an error occurred.", e);
+      }
+    }
+  }
+
   public UpgradeStatus getStatus() {
     return status;
+  }
+
+  /**
+   * Exit loudly if there are outstanding Fate operations. Since Fate serializes class names, we
+   * need to make sure there are no queued transactions from a previous version before continuing an
+   * upgrade. The status of the operations is irrelevant; those in SUCCESSFUL status cause the same
+   * problem as those just queued.
+   *
+   * Note that the Manager should not allow write access to Fate until after all upgrade steps are
+   * complete.
+   *
+   * Should be called as a guard before performing any upgrade steps, after determining that an
+   * upgrade is needed.
+   *
+   * see ACCUMULO-2519
+   */
+  @SuppressFBWarnings(value = "DM_EXIT",
+      justification = "Want to immediately stop all manager threads on upgrade error")
+  private void abortIfFateTransactions(ServerContext context) {
+    try {
+      final ReadOnlyTStore<UpgradeCoordinator> fate =
+          new ReadOnlyStore<>(new ZooStore<>(context.getZooKeeperRoot() + Constants.ZFATE,
+              context.getZooReaderWriter()));
+      if (!(fate.list().isEmpty())) {
+        throw new AccumuloException("Aborting upgrade because there are"
+            + " outstanding FATE transactions from a previous Accumulo version."
+            + " You can start the tservers and then use the shell to delete completed "
+            + " transactions. If there are uncomplete transactions, you will need to roll"
+            + " back and fix those issues. Please see the following page for more information: "
+            + " https://accumulo.apache.org/docs/2.x/troubleshooting/advanced#upgrade-issues");
+      }
+    } catch (Exception exception) {
+      log.error("Problem verifying Fate readiness", exception);
+      System.exit(1);
+    }
   }
 }
