+++ server/manager/src/test/java/org/apache/accumulo/manager/upgrade/AccumuloTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.accumulo.server;
+package org.apache.accumulo.manager.upgrade;
 
 import static org.easymock.EasyMock.createMock;
 import static org.easymock.EasyMock.expect;
@@ -25,8 +25,11 @@ import static org.junit.Assert.assertEquals;
 
 import java.io.FileNotFoundException;
 
+import org.apache.accumulo.core.conf.DefaultConfiguration;
 import org.apache.accumulo.core.volume.Volume;
+import org.apache.accumulo.server.ServerConstants;
 import org.apache.accumulo.server.fs.VolumeManager;
+import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FSDataOutputStream;
 import org.apache.hadoop.fs.FileStatus;
 import org.apache.hadoop.fs.FileSystem;
@@ -39,11 +42,13 @@ import com.google.common.collect.Sets;
 public class AccumuloTest {
   private FileSystem fs;
   private Path path;
+  private ServerConstants serverConstants;
 
   @Before
   public void setUp() {
     fs = createMock(FileSystem.class);
     path = createMock(Path.class);
+    serverConstants = new ServerConstants(DefaultConfiguration.getInstance(), new Configuration());
   }
 
   private FileStatus[] mockPersistentVersion(String s) {
@@ -63,7 +68,7 @@ public class AccumuloTest {
     expect(fs.listStatus(path)).andReturn(files);
     replay(fs);
 
-    assertEquals(42, ServerUtil.getAccumuloPersistentVersion(fs, path));
+    assertEquals(42, serverConstants.getAccumuloPersistentVersion(fs, path));
   }
 
   @Test
@@ -71,7 +76,7 @@ public class AccumuloTest {
     expect(fs.listStatus(path)).andReturn(null);
     replay(fs);
 
-    assertEquals(-1, ServerUtil.getAccumuloPersistentVersion(fs, path));
+    assertEquals(-1, serverConstants.getAccumuloPersistentVersion(fs, path));
   }
 
   @Test
@@ -79,7 +84,7 @@ public class AccumuloTest {
     expect(fs.listStatus(path)).andReturn(new FileStatus[0]);
     replay(fs);
 
-    assertEquals(-1, ServerUtil.getAccumuloPersistentVersion(fs, path));
+    assertEquals(-1, serverConstants.getAccumuloPersistentVersion(fs, path));
   }
 
   @Test(expected = RuntimeException.class)
@@ -87,7 +92,7 @@ public class AccumuloTest {
     expect(fs.listStatus(path)).andThrow(new FileNotFoundException());
     replay(fs);
 
-    assertEquals(-1, ServerUtil.getAccumuloPersistentVersion(fs, path));
+    assertEquals(-1, serverConstants.getAccumuloPersistentVersion(fs, path));
   }
 
   @Test
@@ -96,8 +101,7 @@ public class AccumuloTest {
     FileSystem fs1 = createMock(FileSystem.class);
     Path baseVersion1 = new Path("hdfs://volume1/accumulo/version");
     Path oldVersion1 = new Path("hdfs://volume1/accumulo/version/7");
-    Path newVersion1 = new Path(
-        "hdfs://volume1/accumulo/version/" + Integer.toString(ServerConstants.DATA_VERSION));
+    Path newVersion1 = new Path("hdfs://volume1/accumulo/version/" + ServerConstants.DATA_VERSION);
 
     FileStatus[] files1 = mockPersistentVersion("7");
     expect(fs1.listStatus(baseVersion1)).andReturn(files1);
@@ -114,8 +118,7 @@ public class AccumuloTest {
     FileSystem fs2 = createMock(FileSystem.class);
     Path baseVersion2 = new Path("hdfs://volume2/accumulo/version");
     Path oldVersion2 = new Path("hdfs://volume2/accumulo/version/7");
-    Path newVersion2 = new Path(
-        "hdfs://volume2/accumulo/version/" + Integer.toString(ServerConstants.DATA_VERSION));
+    Path newVersion2 = new Path("hdfs://volume2/accumulo/version/" + ServerConstants.DATA_VERSION);
 
     FileStatus[] files2 = mockPersistentVersion("7");
     expect(fs2.listStatus(baseVersion2)).andReturn(files2);
@@ -136,6 +139,9 @@ public class AccumuloTest {
     expect(vm.create(newVersion2)).andReturn(fsdos2);
     replay(vm);
 
-    ServerUtil.updateAccumuloVersion(vm, 7);
+    UpgradeCoordinator upgradeCoordinator = new UpgradeCoordinator();
+    ServerConstants constants =
+        new ServerConstants(DefaultConfiguration.getInstance(), new Configuration());
+    upgradeCoordinator.updateAccumuloVersion(constants, vm, 7);
   }
 }
